requirements.txt
pymssql>=2.3.0
watchdog>=2.1.0
patool>=1.12
requests>=2.31.0
python-dotenv>=1.0.0
python-json-logger>=2.0.7
pydantic>=2.0.0
pydantic-settings>=2.0.0
backoff>=2.2.1